requirements.txt
autopep8==1.5.7
blinker==1.4
cachelib==0.1.1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==3.7.3
itsdangerous==1.1.0
Jinja2==2.11.3
jira==3.1.1
ldap3==2.9
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
pyasn1==0.4.8
pycodestyle==2.7.0
PyJWT==1.4.2
six==1.15.0
SQLAlchemy==1.4.2
tk==0.1.0
toml==0.10.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1